Fix most warnings reported by msvc
Note that currently we build with low warning level, so we don't get much of them.

There are still security issues with some functions, not sure how if fixing them is portable.

+++ src/mirrors/s3.cpp
@@ -162,7 +162,7 @@ namespace powerloader
         unsigned int DateKeyLen;
         DateKey = HMAC(EVP_sha256(),
                        key1.c_str(),
-                       key1.size(),
+                       static_cast<int>(key1.size()),
                        reinterpret_cast<const unsigned char*>(yyyymmdd.c_str()),
                        yyyymmdd.size(),
                        NULL,
